+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Breaking: Rename `GetPrintableOptions` to `ConfigGetPrintableOptions` and fix.
+- Add exports field to 'package.json'.
+- Add commander.js integration.
+- Add `failString` method.
+- Add `ConfigPair` type.
+- Add `Types.dateString`.
+- Fix typings. Allow `OptionalType<undefined>`. Allow non-string keys: number and symbol.
+- Fix parser option.
+- Improve array type name: `0|1|2[] -> array<0|1|2>`.
+- Improve object type name: `object -> object<0|1|2>`.
+
 ## [3.0.0] - 2024-11-22
 
 - Major changes: